Refactor functionality into discrete components
