+++ src/NextJs.ts
@@ -1,32 +1,20 @@
 import * as path from 'path';
-import { CfnOutput, CustomResource, Duration, Fn, RemovalPolicy, Stack, Token } from 'aws-cdk-lib';
+import { CfnOutput, Duration, Fn, Stack } from 'aws-cdk-lib';
 import * as acm from 'aws-cdk-lib/aws-certificatemanager';
 import * as cloudfront from 'aws-cdk-lib/aws-cloudfront';
 import * as origins from 'aws-cdk-lib/aws-cloudfront-origins';
 import * as iam from 'aws-cdk-lib/aws-iam';
 import * as lambda from 'aws-cdk-lib/aws-lambda';
-import { FunctionProps, LayerVersion } from 'aws-cdk-lib/aws-lambda';
-import { NodejsFunction } from 'aws-cdk-lib/aws-lambda-nodejs';
 import * as route53 from 'aws-cdk-lib/aws-route53';
 import * as route53Patterns from 'aws-cdk-lib/aws-route53-patterns';
 import * as route53Targets from 'aws-cdk-lib/aws-route53-targets';
 import * as s3 from 'aws-cdk-lib/aws-s3';
-import * as s3Assets from 'aws-cdk-lib/aws-s3-assets';
-import { BucketDeployment, Source } from 'aws-cdk-lib/aws-s3-deployment';
-import * as cr from 'aws-cdk-lib/custom-resources';
 import { Construct } from 'constructs';
-
-import * as spawn from 'cross-spawn';
-import * as esbuild from 'esbuild';
 import * as fs from 'fs-extra';
-import * as micromatch from 'micromatch';
-import {
-  BaseSiteCdkDistributionProps,
-  BaseSiteDomainProps,
-  BaseSiteReplaceProps,
-  buildErrorResponsesForRedirectToIndex,
-  getBuildCmdEnvironment,
-} from './BaseSite';
+import { BaseSiteCdkDistributionProps, BaseSiteDomainProps, buildErrorResponsesForRedirectToIndex } from './BaseSite';
+import { NextJsAssetsDeployment } from './NextjsAssetsDeployment';
+import { NextjsBaseProps, NextjsBuild } from './NextjsBuild';
+import { NextJsLambda } from './NextjsLambda';
 // import { isCDKConstruct } from './Construct';
 // import { App } from './App.js';
 // import {
@@ -43,19 +31,13 @@ import {
 
 // const __dirname = url.fileURLToPath(new URL('.', import.meta.url));
 
-const NEXTJS_BUILD_DIR = '.next';
-const NEXTJS_STATIC_DIR = 'static';
-const NEXTJS_PUBLIC_DIR = 'public';
-const NEXTJS_BUILD_STANDALONE_DIR = 'standalone';
-const NEXTJS_BUILD_STANDALONE_ENV = 'NEXT_PRIVATE_STANDALONE';
-
 // contains server-side resolved environment vars in config bucket
-const CONFIG_ENV_JSON_PATH = 'next-env.json';
+// const CONFIG_ENV_JSON_PATH = 'next-env.json';
 
-export interface NextJsDomainProps extends BaseSiteDomainProps {}
-export interface NextJsCdkDistributionProps extends BaseSiteCdkDistributionProps {}
+export interface NextjsDomainProps extends BaseSiteDomainProps {}
+export interface NextjsCdkDistributionProps extends BaseSiteCdkDistributionProps {}
 
-export interface NextJsCachePolicyProps {
+export interface NextjsCachePolicyProps {
   readonly staticCachePolicy?: cloudfront.ICachePolicy;
   readonly lambdaCachePolicy?: cloudfront.ICachePolicy;
   /**
@@ -67,36 +49,28 @@ export interface NextJsCachePolicyProps {
 /**
  * Resources that will be created automatically if not supplied.
  */
-export interface NextJsCdkProps {
+export interface NextjsCdkProps {
   readonly bucket?: s3.BucketProps | s3.IBucket;
   /**
    * Pass in a value to override the default settings this construct uses to
    * create the CDK `Distribution` internally.
    */
-  readonly distribution?: NextJsCdkDistributionProps;
+  readonly distribution?: NextjsCdkDistributionProps;
   /**
    * Override the default CloudFront cache policies created internally.
    */
-  readonly cachePolicies?: NextJsCachePolicyProps;
+  readonly cachePolicies?: NextjsCachePolicyProps;
   /**
    * Override the default CloudFront image origin request policy created internally
    */
   readonly imageOriginRequestPolicy?: cloudfront.IOriginRequestPolicy;
 }
 
-export interface NextJsProps {
+export interface NextjsProps extends NextjsBaseProps {
   /**
    * Allows you to override default settings this construct uses internally to ceate the bucket
    */
-  readonly cdk?: NextJsCdkProps;
-
-  /**
-   * Path to the directory where the NextJS project is located.
-   * Can be the root of your project or a subdirectory.
-   * Preferably an absolute path.
-   */
-  readonly path: string;
-
+  readonly cdk?: NextjsCdkProps;
   /**
    * The customDomain for this website. Supports domains that are hosted
    * either on [Route 53](https://aws.amazon.com/route53/) or externally.
@@ -105,12 +79,12 @@ export interface NextJsProps {
    * [following this guide](https://docs.aws.amazon.com/Route53/latest/DeveloperGuide/MigratingDNS.html).
    *
    * @example
-   * new NextJs(stack, "Site", {
+   * new Nextjs(stack, "Site", {
    *   path: path.resolve("packages/web"),
    *   customDomain: "domain.com",
    * });
    *
-   * new NextJs(stack, "Site", {
+   * new Nextjs(stack, "Site", {
    *   path: path.resolve("packages/web")",
    *   customDomain: {
    *     domainName: "domain.com",
@@ -119,59 +93,23 @@ export interface NextJsProps {
    *   },
    * });
    */
-  readonly customDomain?: string | NextJsDomainProps;
-
-  /**
-   * An object with the key being the environment variable name.
-   *
-   * @example
-   * new NextJs(stack, "Site", {
-   *   path: path.resolve("packages/web"),
-   *   environment: {
-   *     API_URL: api.url,
-   *     NEXT_PUBLIC_USER_POOL_CLIENT_ID: auth.cognitoUserPoolClient.userPoolClientId,
-   *   },
-   * });
-   */
-  readonly environment?: Record<string, string>;
-
-  /**
-   * Optional value for NODE_ENV during build and runtime.
-   */
-  readonly nodeEnv?: string;
-
-  /**
-   * Defaults for lambda functions created by this construct.
-   */
-  readonly functionDefaults?: FunctionProps;
+  readonly customDomain?: string | NextjsDomainProps;
 
   /**
    * While deploying, waits for the CloudFront cache invalidation process to finish. This ensures that the new content will be served once the deploy command finishes. However, this process can sometimes take more than 5 mins. For non-prod environments it might make sense to pass in `false`. That'll skip waiting for the cache to invalidate and speed up the deploy process.
    */
   readonly waitForInvalidation?: boolean;
-
-  /**
-   * Skip building app and deploy a placeholder.
-   * Useful when using `next dev` for local development.
-   */
-  readonly isPlaceholder: boolean;
-
-  /**
-   * Directory to store temporary build files in.
-   * Defaults to os.mkdtempSync().
-   */
-  readonly tempBuildDir?: string;
 }
 
 /**
- * The `NextJs` construct is a higher level CDK construct that makes it easy to create a Nextjs app.
+ * The `Nextjs` construct is a higher level CDK construct that makes it easy to create a Nextjs app.
  *
  * Your standalone application will be bundled using output tracing and will be deployed to a Lambda function.
  * You must use Next.js 10.3.0 or newer.
  *
 
  */
-export class NextJs extends Construct {
+export class Nextjs extends Construct {
   /**
    * The default CloudFront cache policy properties for static pages.
    */
@@ -199,7 +137,7 @@ export class NextJs extends Construct {
     minTtl: Duration.days(0),
     enableAcceptEncodingBrotli: true,
     enableAcceptEncodingGzip: true,
-    comment: 'NextJs Image Default Cache Policy',
+    comment: 'Nextjs Image Default Cache Policy',
   };
 
   /**
@@ -214,7 +152,7 @@ export class NextJs extends Construct {
     minTtl: Duration.seconds(0),
     enableAcceptEncodingBrotli: true,
     enableAcceptEncodingGzip: true,
-    comment: 'NextJs Lambda Default Cache Policy',
+    comment: 'Nextjs Lambda Default Cache Policy',
   };
 
   /**
@@ -222,23 +160,29 @@ export class NextJs extends Construct {
    */
   public static imageOriginRequestPolicyProps: cloudfront.OriginRequestPolicyProps = {
     queryStringBehavior: cloudfront.OriginRequestQueryStringBehavior.all(),
-    comment: 'NextJs Lambda Default Origin Request Policy',
+    comment: 'Nextjs Lambda Default Origin Request Policy',
   };
 
-  // files to rewrite CloudFormation tokens in environment variables
-  private replaceTokenGlobs = ['**/*.html', '**/*.js', '**/*.cjs', '**/*.mjs', '**/*.json'];
+  /**
+   * The main NextJS server handler lambda function.
+   */
+  public serverFunction: NextJsLambda;
 
   /**
-   * Exposes CDK instances created within the construct.
+   * Built NextJS project output.
    */
+  public build: NextjsBuild;
+
   /**
-   * The main Nextjs server handler lambda function.
+   * Bucket containing NextJS static assets.
    */
-  serverFunction: lambda.Function;
+  public bucket: s3.IBucket;
+
   /**
-   * The internally created CDK `Bucket` instance.
+   * Asset deployment to S3.
    */
-  bucket: s3.Bucket;
+  public assetsDeployment: NextJsAssetsDeployment;
+
   /**
    * The internally created CDK `Distribution` instance.
    */
@@ -252,24 +196,14 @@ export class NextJs extends Construct {
    */
   certificate?: acm.ICertificate;
 
-  private props: NextJsProps;
-  /**
-   * Skip building app and deploy a placeholder.
-   * Useful when using `next dev` for local development.
-   */
-  public isPlaceholder: boolean;
-  /**
-   * The root directory used for builds.
-   */
-  private tempBuildDir: string;
-  // private awsCliLayer: AwsCliLayer;
+  protected props: NextjsProps;
+
   public originAccessIdentity: cloudfront.IOriginAccessIdentity;
-  configBucket?: s3.Bucket;
+  public tempBuildDir: string;
+  public configBucket?: s3.Bucket;
 
-  constructor(scope: Construct, id: string, props: NextJsProps) {
+  constructor(scope: Construct, id: string, props: NextjsProps) {
     super(scope, id);
-
-    this.isPlaceholder = props.isPlaceholder;
 
     // get dir to store temp build files in
     this.tempBuildDir = props.tempBuildDir
@@ -279,20 +213,22 @@ export class NextJs extends Construct {
     this.props = props;
     // this.awsCliLayer = new AwsCliLayer(this, 'AwsCliLayer');
     this.registerSiteEnvironment();
-    this.configBucket = this.createConfigBucket();
+    // this.configBucket = this.createConfigBucket();
 
-    // Prepare app
-    if (this.isPlaceholder) {
-    } else {
-      this.buildApp();
-    }
+    // build nextjs app
+    this.build = new NextjsBuild(scope, id, props);
+    this.serverFunction = new NextJsLambda(this, 'Fn', { ...props, build: this.build });
+    this.assetsDeployment = new NextJsAssetsDeployment(this, 'AssetDeployment', {
+      ...props,
+      build: this.build,
+    });
+    this.bucket = this.assetsDeployment.bucket;
 
     this.originAccessIdentity = new cloudfront.OriginAccessIdentity(this, 'OAI(', {
       comment: 'Allows CloudFront to access S3 bucket with assets',
     });
 
-    // Create Bucket which will be utilised to contain the statics
-    this.bucket = this.createAssetBucket();
+    // allow cloudfront to access assets bucket
     this.bucket.addToResourcePolicy(
       new iam.PolicyStatement({
         // only allow getting of files - not listing
@@ -301,25 +237,19 @@ export class NextJs extends Construct {
         principals: [this.originAccessIdentity.grantPrincipal],
       })
     );
-
-    // Create Server function
-    this.serverFunction = this.createServerFunction();
-
     // Create Custom Domain
     this.validateCustomDomainSettings();
     this.hostedZone = this.lookupHostedZone();
     this.certificate = this.createCertificate();
 
-    // Create S3 Deployment
-    const s3Deployments = this.uploadS3Assets();
-
     // Create CloudFront
-    this.distribution = this.isPlaceholder
+    this.distribution = this.props.isPlaceholder
       ? this.createCloudFrontDistributionForStub()
       : this.createCloudFrontDistribution();
-    s3Deployments.forEach((s3Deployment) => this.distribution.node.addDependency(s3Deployment));
+    // wait for asset deployments to finish
+    this.assetsDeployment.deployments.forEach((s3Deployment) => this.distribution.node.addDependency(s3Deployment));
 
-    // // Invalidate CloudFront (might already be handled by S3 deployment?)
+    // // Invalidate CloudFront (might already be handled by deployments?)
     // const invalidationCR = this.createCloudFrontInvalidation();
     // invalidationCR.node.addDependency(this.distribution);
 
@@ -390,535 +320,13 @@ export class NextJs extends Construct {
   // FIXME: SST
   // public getConstructMetadata() {
   //   return {
-  //     type: 'NextJs' as const,
+  //     type: 'Nextjs' as const,
   //     data: {
   //       distributionId: this.distribution.distributionId,
   //       customDomainUrl: this.customDomainUrl,
   //     },
   //   };
   // }
-
-  /////////////////////
-  // Build App
-  /////////////////////
-
-  private buildApp() {
-    // Build
-    this.runNpmBuild();
-
-    // Validate build output exists
-    const baseOutputDir = path.resolve(this.props.path);
-    const serverBuildDir = path.join(baseOutputDir, NEXTJS_BUILD_DIR);
-
-    if (!fs.existsSync(serverBuildDir)) {
-      throw new Error(`No server build output found at "${serverBuildDir}"`);
-    }
-  }
-
-  private runNpmBuild() {
-    const { path: sitePath } = this.props;
-
-    // validate site path exists
-    if (!fs.existsSync(sitePath)) {
-      throw new Error(`No path found at "${path.resolve(sitePath)}"`);
-    }
-
-    // Ensure that the site has a build script defined
-    if (!fs.existsSync(path.join(sitePath, 'package.json'))) {
-      throw new Error(`No package.json found at "${sitePath}".`);
-    }
-    const packageJson = fs.readJsonSync(path.join(sitePath, 'package.json'));
-    if (!packageJson.scripts || !packageJson.scripts.build) {
-      throw new Error(`No "build" script found within package.json in "${sitePath}".`);
-    }
-
-    // Run build
-    console.debug('Running "npm build" script');
-    const buildEnv = {
-      ...process.env,
-      [NEXTJS_BUILD_STANDALONE_ENV]: 'true',
-      ...(this.props.nodeEnv ? { NODE_ENV: this.props.nodeEnv } : {}),
-      ...getBuildCmdEnvironment(this.props.environment),
-    };
-    const buildResult = spawn.sync('npm', ['run', 'build'], {
-      cwd: sitePath,
-      stdio: 'inherit',
-      env: buildEnv,
-    });
-    if (buildResult.status !== 0) {
-      throw new Error('The app "build" script failed.');
-    }
-  }
-
-  private buildLayer(): LayerVersion {
-    const layerDir = path.resolve(__dirname, '../assets');
-    const sharpLayer = new LayerVersion(this, 'SharpLayer', {
-      code: new lambda.AssetCode(path.join(layerDir, 'sharp-0.30.0.zip')),
-      compatibleRuntimes: [lambda.Runtime.NODEJS_16_X],
-      description: 'Sharp 0.30.0 for NextJS',
-    });
-    return sharpLayer;
-
-    ///////  other ways to build this layer:
-    // const buildDir = path.resolve(
-    //   path.join(this.sstBuildDir, `NextjsLayer-${this.node.id}-${this.node.addr}`)
-    // );
-    // fs.removeSync(buildDir);
-    // fs.mkdirSync(buildDir, { recursive: true });
-    // const zipFile ="nextjs-layer.zip"
-    // const zipFilePath = path.join(buildDir, zipFile);
-    // const LAMBDA_FOLDER = 'nodejs'
-    // const createBundleCmdArgs = [
-    //   '-xc',
-    //   [
-    //     `mkdir -p ${LAMBDA_FOLDER}`,
-    //     `cd ${LAMBDA_FOLDER}`,
-    //     `npm install \
-    //     --arch=x64 \
-    //     --platform=linux \
-    //     --target=16.15 \
-    //     --libc=glibc \
-    //     next sharp`,
-    //     'cd ..',
-    //     `zip -qr ${zipFile} ${LAMBDA_FOLDER}`
-    //   ].join(' && '),
-    // ];
-
-    // const buildResult = spawn.sync('bash', createBundleCmdArgs, {
-    //   cwd: buildDir,
-    //   stdio: "inherit",
-    // });
-    // if (buildResult.status !== 0 || !fs.existsSync(zipFilePath)) {
-    //   throw new Error(`Failed to create nextjs layer in ${buildDir}`);
-    // }
-
-    // // hash our parameters so we know when we need t rebuild
-    // const bundleCommandHash = crypto.createHash('sha256');
-    // bundleCommandHash.update(JSON.stringify(createBundleCmdArgs));
-
-    // // bundle
-    // const code = Code.fromAsset(zipFilePath);
-
-    // // const code = Code.fromAsset(__dirname, {
-    // //   // don't send all our files to docker (slow)
-    // //   ignoreMode: IgnoreMode.GLOB,
-    // //   exclude: ['*'],
-
-    // //   // if our bundle commands (basically our "dockerfile") changes then rebuild the image
-    // //   assetHashType: AssetHashType.CUSTOM,
-    // //   assetHash: bundleCommandHash.digest('hex'),
-
-    // //   bundling: {
-    // //     image: lambda.Runtime.NODEJS_16_X.bundlingImage,
-    // //     command: createBundleCommand,
-    // //   },
-    // // });
-
-    // // Build Next.js layer
-    // const nextjsLayer = new lambda.LayerVersion(this, "NextjsLayer", {
-    //   code,
-    //   compatibleRuntimes: [lambda.Runtime.NODEJS_16_X],
-    //   description: "Next.js",
-    // });
-    // return nextjsLayer;
-  }
-
-  /////////////////////
-  // Bundle S3 Assets
-  /////////////////////
-
-  private uploadS3Assets() {
-    const deployments: BucketDeployment[] = [];
-
-    // path to public folder; root static assets
-    const staticDir = this._getNextStaticDir();
-    let publicDir = this.isPlaceholder
-      ? path.resolve(__dirname, '../assets/placeholder-site')
-      : this._getNextPublicDir();
-
-    // static dir
-    if (!this.isPlaceholder && fs.existsSync(staticDir)) {
-      // upload static assets
-      deployments.push(
-        new BucketDeployment(this, 'StaticAssetsDeployment', {
-          destinationBucket: this.bucket,
-          destinationKeyPrefix: '_next/static',
-          sources: [Source.asset(staticDir)],
-          distribution: this.distribution, // invalidate Cloudfront distribution caches
-          prune: false, // do not delete stale files
-        })
-      );
-    }
-
-    // public dir
-    if (fs.existsSync(publicDir)) {
-      // zip up assets
-      const zipFilePath = this.createArchive(publicDir, 'public.zip', '*');
-
-      // upload public files to root of S3 bucket
-      deployments.push(
-        new BucketDeployment(this, 'PublicFilesDeployment', {
-          destinationBucket: this.bucket,
-          destinationKeyPrefix: this.isPlaceholder ? '/placeholder' : '/',
-          sources: [Source.asset(zipFilePath)],
-          distribution: this.distribution,
-          prune: false,
-        })
-      );
-    }
-
-    // do rewrites of unresolved CDK tokens in static files
-    const rewriter = this.createRewriteResource();
-    rewriter?.node.addDependency(...deployments);
-
-    return deployments;
-  }
-
-  private createRewriteResource() {
-    const s3keys = this._getStaticFilesForRewrite();
-    if (s3keys.length === 0) return;
-
-    // create a custom resource to find and replace tokenized strings in static files
-    // must happen after deployment when tokens can be resolved
-    const rewriteFn = new lambda.Function(this, 'OnEventHandler', {
-      runtime: lambda.Runtime.NODEJS_16_X,
-      memorySize: 1024,
-      handler: 'index.handler',
-      code: lambda.Code.fromInline(`
-      const AWS = require('aws-sdk');
-
-      // search and replace tokenized values of designated objects in s3
-      exports.handler = async (event) => {
-        const requestType = event.RequestType;
-        if (requestType === 'Create' || requestType === 'Update') {
-          // rewrite static files
-          const s3 = new AWS.S3();
-          const {s3keys, bucket, replacements} = event.ResourceProperties;
-          if (!s3keys || !bucket || !replacements) {
-            console.error("Missing required properties")
-            return
-          }
-          const promises = s3keys.map(async (key) => {
-            const params = {Bucket: bucket, Key: key};
-            const data = await s3.getObject(params).promise();
-            let body = data.Body.toString('utf-8');
-
-            // do replacements of tokens
-            replacements.forEach((replacement) => {
-              body = body.replace(replacement.search, replacement.replace);
-            });
-            const putParams = {
-              ...params,
-              Body: body,
-              ContentType: data.ContentType,
-              ContentEncoding: data.ContentEncoding,
-              CacheControl: data.CacheControl,
-            }
-            await s3.putObject(putParams).promise();
-          });
-          await Promise.all(promises);
-        }
-
-        return event;
-      };
-      `),
-      initialPolicy: [
-        new iam.PolicyStatement({
-          actions: ['s3:GetObject', 's3:PutObject'],
-          resources: [this.bucket.arnForObjects('*')],
-        }),
-      ],
-    });
-
-    // custom resource to run the rewriter after files are copied and we can resolve token values
-    const provider = new cr.Provider(this, 'RewriteStaticProvider', {
-      onEventHandler: rewriteFn,
-    });
-    return new CustomResource(this, 'RewriteStatic', {
-      serviceToken: provider.serviceToken,
-      properties: {
-        bucket: this.bucket.bucketName,
-        s3keys,
-        replacements: this._getS3ContentReplaceValues(),
-      },
-    });
-  }
-
-  private _getStaticFilesForRewrite() {
-    const s3keys: string[] = [];
-    const replaceValues = this._getS3ContentReplaceValues();
-    if (!replaceValues) return s3keys;
-
-    // where to find static files
-    const globs = [...new Set(replaceValues.flatMap((replaceValue) => replaceValue.files))];
-    const searchDirs = [
-      { dir: this._getNextStaticDir(), prefix: '_next/static' },
-      { dir: this._getNextPublicDir(), prefix: '' },
-    ];
-
-    // traverse static dirs
-    searchDirs.forEach(({ dir, prefix }) => {
-      if (!fs.existsSync(dir)) {
-        return;
-      }
-      this.listDirectory(dir).forEach((file) => {
-        const relativePath = path.relative(dir, file);
-
-        // is this file a glob match?
-        if (!micromatch.isMatch(relativePath, globs, { dot: true })) {
-          return;
-        }
-        s3keys.push(`${prefix}/${relativePath}`);
-      });
-    });
-    return s3keys;
-  }
-
-  /*
-  private createLambdaCodeReplacer(name: string, asset: s3Assets.Asset): CustomResource {
-    // Note: Source code for the Lambda functions have "{{ ENV_KEY }}" in them.
-    //       They need to be replaced with real values before the Lambda
-    //       functions get deployed.
-
-    const providerId = 'LambdaCodeReplacerProvider';
-    const resId = `${name}LambdaCodeReplacer`;
-    const stack = Stack.of(this);
-    let provider = stack.node.tryFindChild(providerId) as lambda.Function;
-
-    // Create provider if not already created
-    if (!provider) {
-      provider = new lambda.Function(stack, providerId, {
-        code: lambda.Code.fromAsset(path.join(__dirname, '../assets/NextjsSite/custom-resource')),
-        layers: [this.awsCliLayer],
-        runtime: lambda.Runtime.PYTHON_3_7,
-        handler: 'lambda-code-updater.handler',
-        timeout: Duration.minutes(15),
-        memorySize: 1024,
-      });
-    }
-
-    // Allow provider to perform search/replace on the asset
-    provider.role?.addToPrincipalPolicy(
-      new iam.PolicyStatement({
-        effect: iam.Effect.ALLOW,
-        actions: ['s3:*'],
-        resources: [`arn:aws:s3:::${asset.s3BucketName}/${asset.s3ObjectKey}`],
-      })
-    );
-
-    // Create custom resource
-    const resource = new CustomResource(this, resId, {
-      serviceToken: provider.functionArn,
-      resourceType: 'Custom::SSTLambdaCodeUpdater',
-      properties: {
-        Source: {
-          BucketName: asset.s3BucketName,
-          ObjectKey: asset.s3ObjectKey,
-        },
-        ReplaceValues: this._getLambdaContentReplaceValues(),
-      },
-    });
-
-    return resource;
-  }*/
-
-  private createAssetBucket(): s3.Bucket {
-    const { cdk } = this.props;
-
-    // cdk.bucket is an imported construct
-    if (cdk?.bucket && isCDKConstruct(cdk?.bucket)) {
-      return cdk.bucket as s3.Bucket;
-    }
-    // cdk.bucket is a prop
-    else {
-      const bucketProps = cdk?.bucket as s3.BucketProps;
-      return new s3.Bucket(this, 'S3Bucket', {
-        publicReadAccess: true,
-        autoDeleteObjects: true,
-        removalPolicy: RemovalPolicy.DESTROY,
-        ...bucketProps,
-      });
-    }
-  }
-
-  // this can hold our resolved environment vars for the server
-  private createConfigBucket() {
-    // won't work until this is fixed: https://github.com/aws/aws-cdk/issues/19257
-    return undefined;
-    // const bucket = new s3.Bucket(this, "ConfigBucket", { removalPolicy: RemovalPolicy.DESTROY, });
-    // upload environment config to s3
-    // new BucketDeployment(this, 'EnvJsonDeployment', {
-    //   sources: [Source.jsonData(CONFIG_ENV_JSON_PATH, this.props.environment)],
-    //   destinationBucket: bucket,
-    // })
-    // return bucket
-  }
-
-  // zip up a directory and return path to zip file
-  private createArchive(directory: string, zipFileName: string, fileGlob: string = '*'): string {
-    // get output path
-    const zipOutDir = path.resolve(path.join(this.tempBuildDir, `standalone-${this.node.id}-${this.node.addr}`));
-    fs.removeSync(zipOutDir);
-    fs.mkdirpSync(zipOutDir);
-    const zipFilePath = path.join(zipOutDir, zipFileName);
-
-    // run script to create zipfile, preserving symlinks for node_modules (e.g. pnpm structure)
-    const result = spawn.sync(
-      'bash', // getting ENOENT when specifying 'node' here for some reason
-      ['-xc', [`cd '${directory}'`, `zip -ryq2 '${zipFilePath}' ${fileGlob}`].join('&&')],
-      { stdio: 'inherit' }
-    );
-    if (result.status !== 0) {
-      throw new Error(
-        `There was a problem generating the package for ${zipFileName} with ${directory}: ${result.error}`
-      );
-    }
-    // check output
-    if (!fs.existsSync(zipFilePath)) {
-      throw new Error(
-        `There was a problem generating the archive for ${directory}; the archive is missing in ${zipFilePath}.`
-      );
-    }
-
-    return zipFilePath;
-  }
-
-  /////////////////////
-  // Bundle Lambda Server
-  /////////////////////
-
-  private createServerFunction(): NodejsFunction {
-    const { functionDefaults, environment, path: nextjsPath } = this.props;
-
-    // build native deps layer
-    const nextLayer = this.buildLayer();
-
-    // build and bundle the handler
-    const assetPath = this.createServerAsset();
-    const s3asset = new s3Assets.Asset(this, 'MainFnAsset', {
-      path: assetPath,
-    });
-    const code = this.isPlaceholder
-      ? lambda.Code.fromInline(
-          "module.exports.handler = async () => { return { statusCode: 200, body: 'SST placeholder site' } }"
-        )
-      : lambda.Code.fromBucket(s3asset.bucket, s3asset.s3ObjectKey);
-
-    // build the lambda function
-    const fn = new lambda.Function(this, 'MainFn', {
-      ...functionDefaults,
-      memorySize: functionDefaults?.memorySize || 1024,
-      timeout: functionDefaults?.timeout ?? Duration.seconds(10),
-      runtime: lambda.Runtime.NODEJS_16_X,
-      handler: path.join(nextjsPath, 'server.handler'),
-      layers: [nextLayer],
-      code,
-      environment: {
-        ...environment,
-        ...(this.props.nodeEnv ? { NODE_ENV: this.props.nodeEnv } : {}),
-        ...(this.configBucket
-          ? {
-              NEXTJS_SITE_CONFIG_BUCKET_NAME: this.configBucket.bucketName,
-              NEXTJS_SITE_CONFIG_ENV_JSON_PATH: CONFIG_ENV_JSON_PATH,
-            }
-          : {}),
-      },
-    });
-    this.serverFunction = fn;
-
-    // attach permissions
-    this.bucket.grantReadWrite(fn.role!);
-    this.configBucket?.grantRead(fn.role!);
-
-    // FIXME: SST
-    // if (defaults?.function?.permissions) {
-    //   attachPermissionsToRole(fn.role as iam.Role, defaults.function.permissions);
-    // }
-
-    // resolve lambda env vars after deployment
-    if (!this.isPlaceholder && assetPath) {
-      // TODO: only needed if we want to deploy as an edge function because
-      // they don't support environment variables
-      // const updaterCR = this.createLambdaCodeReplacer('MainFn', s3asset);
-      // fn.node.addDependency(updaterCR);
-    }
-
-    return fn;
-  }
-
-  private bundleServerHandler() {
-    const { path: nextjsPath } = this.props;
-
-    const standaloneDirAbsolute = this._getNextStandaloneDir();
-    // delete default nextjs handler if it exists
-    const defaultServerPath = path.join(standaloneDirAbsolute, nextjsPath, 'server.js');
-    if (fs.existsSync(defaultServerPath)) {
-      fs.unlinkSync(defaultServerPath);
-    }
-
-    // rewrite env var placeholders
-    this.rewriteEnvVars();
-
-    // build our server handler
-    const serverHandler = path.resolve(__dirname, './NextJsHandler.ts');
-    // server should live in the same dir as the nextjs app to access deps properly
-    const serverPath = path.join(nextjsPath, 'server.cjs');
-    const esbuildResult = esbuild.buildSync({
-      entryPoints: [serverHandler],
-      bundle: true,
-      minify: false,
-      sourcemap: true,
-      target: 'node16',
-      platform: 'node',
-      external: ['sharp', 'next'],
-      format: 'cjs',
-      outfile: path.join(standaloneDirAbsolute, serverPath),
-    });
-    if (esbuildResult.errors.length > 0) {
-      esbuildResult.errors.forEach((error) => console.error(error));
-      throw new Error('There was a problem bundling the server.');
-    }
-  }
-
-  private rewriteEnvVars() {
-    // undo inlining of NEXT_PUBLIC_ env vars for server code
-    // https://github.com/vercel/next.js/issues/40827
-    const replaceValues = this._getServerContentReplaceValues();
-    const globs = this.replaceTokenGlobs;
-
-    // traverse server dirs
-    const searchDir = this._getNextStandaloneBuildDir();
-    if (!fs.existsSync(searchDir)) return;
-
-    this.listDirectory(searchDir).forEach((file) => {
-      const relativePath = path.relative(searchDir, file);
-      if (!micromatch.isMatch(relativePath, globs, { dot: true })) {
-        return;
-      }
-
-      // matches file search pattern
-      // do replacements
-      let fileContent = fs.readFileSync(file, 'utf8');
-      Object.entries(replaceValues).forEach(([key, value]) => {
-        console.log(`Replacing ${key} with ${value} in ${file}`);
-        fileContent = fileContent.replace(key, value);
-      });
-      fs.writeFileSync(file, fileContent);
-    });
-  }
-
-  // create the lambda code and zip it
-  // returns path to zip file
-  private createServerAsset(): string {
-    const standaloneDirAbsolute = this._getNextStandaloneDir();
-
-    // build our handler
-    this.bundleServerHandler();
-
-    // zip up the directory
-    const zipFilePath = this.createArchive(standaloneDirAbsolute, 'standalone.zip');
-    return zipFilePath;
-  }
 
   /////////////////////
   // CloudFront Distribution
@@ -946,7 +354,7 @@ export class NextJs extends Construct {
     const viewerProtocolPolicy = cloudfront.ViewerProtocolPolicy.REDIRECT_TO_HTTPS;
 
     // handle placeholder
-    if (this.isPlaceholder) {
+    if (this.props.isPlaceholder) {
       return new cloudfront.Distribution(this, 'Distribution', {
         defaultRootObject: 'placeholder/index.html',
         errorResponses: buildErrorResponsesForRedirectToIndex('placeholder/index.html'),
@@ -1006,7 +414,7 @@ export class NextJs extends Construct {
 
     // if we don't have a static file called index.html then we should
     // redirect to the lambda handler
-    const hasIndexHtml = this.readPublicFileList().includes('index.html');
+    const hasIndexHtml = this.build.readPublicFileList().includes('index.html');
 
     return new cloudfront.Distribution(this, 'Distribution', {
       // defaultRootObject: "index.html",
@@ -1058,22 +466,22 @@ export class NextJs extends Construct {
   }
 
   private createCloudFrontStaticCachePolicy(): cloudfront.CachePolicy {
-    return new cloudfront.CachePolicy(this, 'StaticsCache', NextJs.staticCachePolicyProps);
+    return new cloudfront.CachePolicy(this, 'StaticsCache', Nextjs.staticCachePolicyProps);
   }
 
   // TODO
   /*
   private createCloudFrontImageCachePolicy(): cloudfront.CachePolicy {
-    return new cloudfront.CachePolicy(this, 'ImageCache', NextJs.imageCachePolicyProps);
+    return new cloudfront.CachePolicy(this, 'ImageCache', Nextjs.imageCachePolicyProps);
   }
 
   private createCloudFrontImageOriginRequestPolicy(): cloudfront.OriginRequestPolicy {
-    return new cloudfront.OriginRequestPolicy(this, 'ImageOriginRequest', NextJs.imageOriginRequestPolicyProps);
+    return new cloudfront.OriginRequestPolicy(this, 'ImageOriginRequest', Nextjs.imageOriginRequestPolicyProps);
   }
-	*/
+  */
 
   private createCloudFrontLambdaCachePolicy(): cloudfront.CachePolicy {
-    return new cloudfront.CachePolicy(this, 'LambdaCache', NextJs.lambdaCachePolicyProps);
+    return new cloudfront.CachePolicy(this, 'LambdaCache', Nextjs.lambdaCachePolicyProps);
   }
 
   private createCloudFrontDistributionForStub(): cloudfront.Distribution {
@@ -1140,7 +548,7 @@ export class NextJs extends Construct {
       },
     });
   }
-	*/
+  */
 
   /////////////////////
   // Custom Domain
@@ -1303,150 +711,4 @@ export class NextJs extends Construct {
     //   environmentOutputs,
     // } as BaseSiteEnvironmentOutputsInfo);
   }
-
-  private _getS3ContentReplaceValues(): BaseSiteReplaceProps[] {
-    const replaceValues: BaseSiteReplaceProps[] = [];
-
-    Object.entries(this.props.environment || {})
-      .filter(([, value]) => Token.isUnresolved(value))
-      .forEach(([key, value]) => {
-        const token = `{{ ${key} }}`;
-        replaceValues.push(
-          ...this.replaceTokenGlobs.map((glob) => ({
-            files: glob,
-            search: token,
-            replace: value,
-          }))
-        );
-      });
-    return replaceValues;
-  }
-
-  // replace inlined public env vars with calls to process.env
-  private _getServerContentReplaceValues(): Record<string, string> {
-    return Object.fromEntries(
-      Object.entries(this.props.environment || {})
-        .filter(([key]) => key.startsWith('NEXT_PUBLIC_'))
-        .map(([key]) => [`"{{ ${key} }}"`, `process.env.${key}`]) // will need to replace with actual value for edge functions
-    );
-  }
-
-  // TODO: needed for edge function support probably
-  // private _getLambdaContentReplaceValues(): BaseSiteReplaceProps[] {
-  //   const replaceValues: BaseSiteReplaceProps[] = [];
-
-  //   // The Next.js app can have environment variables like
-  //   // `process.env.API_URL` in the JS code. `process.env.API_URL` might or
-  //   // might not get resolved on `next build` if it is used in
-  //   // server-side functions, ie. getServerSideProps().
-  //   // Because Lambda@Edge does not support environment variables, we will
-  //   // use the trick of replacing "{{ _SST_NEXTJS_SITE_ENVIRONMENT_ }}" with
-  //   // a JSON encoded string of all environment key-value pairs. This string
-  //   // will then get decoded at run time.
-  //   const lambdaEnvs: { [key: string]: string } = {};
-
-  //   Object.entries(this.props.environment || {}).forEach(([key, value]) => {
-  //     const token = `{{ ${key} }}`;
-  //     replaceValues.push(
-  //       ...this.replaceTokenGlobs.map((glob) => ({
-  //         files: glob,
-  //         search: token,
-  //         replace: value,
-  //       }))
-  //     );
-  //     lambdaEnvs[key] = value;
-  //   });
-
-  //   replaceValues.push(
-  //     {
-  //       files: '**/*.mjs',
-  //       search: '"{{ _SST_NEXTJS_SITE_ENVIRONMENT_ }}"',
-  //       replace: JSON.stringify(lambdaEnvs),
-  //     },
-  //     {
-  //       files: '**/*.cjs',
-  //       search: '"{{ _SST_NEXTJS_SITE_ENVIRONMENT_ }}"',
-  //       replace: JSON.stringify(lambdaEnvs),
-  //     },
-  //     {
-  //       files: '**/*.js',
-  //       search: '"{{ _SST_NEXTJS_SITE_ENVIRONMENT_ }}"',
-  //       replace: JSON.stringify(lambdaEnvs),
-  //     }
-  //   );
-
-  //   return replaceValues;
-  // }
-
-  // getNextBuildId() {
-  //   return fs.readFileSync(path.join(this._getNextStandaloneBuildDir(), 'BUILD_ID'), 'utf-8');
-  // }
-
-  private listDirectory(dir: string) {
-    const fileList: string[] = [];
-    const publicFiles = fs.readdirSync(dir);
-    for (const filename of publicFiles) {
-      const filepath = path.join(dir, filename);
-      const stat = fs.statSync(filepath);
-      if (stat.isDirectory()) {
-        fileList.push(...this.listDirectory(filepath));
-      } else {
-        fileList.push(filepath);
-      }
-    }
-
-    return fileList;
-  }
-
-  readPublicFileList() {
-    const publicDir = this._getNextPublicDir();
-    return this.listDirectory(publicDir).map((file) => path.join('/', path.relative(publicDir, file)));
-  }
-
-  // get the path to the directory containing the nextjs project
-  // it may be the project root or a subdirectory in a monorepo setup
-  private _getNextDir() {
-    const { path: nextjsPath } = this.props; // path to nextjs dir inside project
-    const absolutePath = path.resolve(nextjsPath); // e.g. /home/me/myapp/web
-    if (!fs.existsSync(absolutePath)) {
-      throw new Error(`Could not find ${absolutePath} directory.`);
-    }
-    return absolutePath;
-  }
-
-  // .next
-  private _getNextBuildDir() {
-    return path.join(this._getNextDir(), NEXTJS_BUILD_DIR);
-  }
-
-  // output of nextjs standalone build
-  private _getNextStandaloneDir() {
-    const nextDir = this._getNextBuildDir();
-    const standaloneDir = path.join(nextDir, NEXTJS_BUILD_STANDALONE_DIR);
-
-    if (!fs.existsSync(standaloneDir)) {
-      throw new Error(`Could not find ${standaloneDir} directory.`);
-    }
-    return standaloneDir;
-  }
-
-  // nextjs project inside of standalone build
-  // contains manifests and server code
-  private _getNextStandaloneBuildDir() {
-    return path.join(this._getNextStandaloneDir(), this.props.path, NEXTJS_BUILD_DIR);
-  }
-
-  // contains static files
-  private _getNextStaticDir() {
-    return path.join(this._getNextBuildDir(), NEXTJS_STATIC_DIR);
-  }
-  private _getNextPublicDir() {
-    return path.join(this._getNextDir(), NEXTJS_PUBLIC_DIR);
-  }
-}
-// taken from https://github.com/serverless-stack/sst/blob/8d377e941467ced81d8cc31ee67d5a06550f04d4/packages/resources/src/Construct.ts
-const JSII_RTTI_SYMBOL_1 = Symbol.for('jsii.rtti');
-function isCDKConstruct(construct: any): construct is Construct {
-  const fqn = construct?.constructor?.[JSII_RTTI_SYMBOL_1]?.fqn;
-  return typeof fqn === 'string' && (fqn.startsWith('@aws-cdk/') || fqn.startsWith('aws-cdk-lib'));
 }
